slpitting componets
